pyproject.toml

[project]
    name = "diffsaverl"
    version = "0.1.0"
    description = "Add your description here"
    requires-python = ">=3.13"
    dependencies = [
        "gymnasium>=1.0.0",
        "wandb>=0.17.7",
        "numpy>=1.26.3",
        "matplotlib>=3.9.2",
        "marimo>=0.8.2",
        "torch>=2.6.0",
        "torchvision>=0.19.0",
        "torchaudio>=2.4.0",
        "pygame>=2.6.0",
        "tornado>=6.4.1",
        "hydra-core>=1.3.2",
        "omegaconf>=2.3.0",
        "moviepy>=1.0.3",
        "tensorboard>=2.17.1",
        "tqdm>=4.66.5",
        "plotly>=5.23.0",
        "optuna>=3.6.1",
        "optuna-integration>=3.6.0",
        "optuna-fast-fanova>=0.0.4",
        "gunicorn>=23.0.0",
        "optuna-dashboard>=0.16.1",
        "jaxtyping>=0.2.34",
        "beartype>=0.18.5",
        "einops>=0.8.0",
        "cvxpy>=1.5.3",
        "diffcp>=1.0.23",
        "cvxpylayers>=0.1.6",
        "pillow>=10.2.0",
        "scipy>=1.14.1",
        "tabulate>=0.9.0",
        "kaleido>=0.2.1",
]
